Moved lib/config -> lib/models/config
