+++ exasol_integration_test_docker_environment/lib/api/build_test_container.py
@@ -13,17 +13,18 @@ from exasol_integration_test_docker_environment.lib.base.run_task import (
     generate_root_task,
     run_task,
 )
-from exasol_integration_test_docker_environment.lib.config.build_config import (
-    set_build_config,
-)
-from exasol_integration_test_docker_environment.lib.config.docker_config import (
-    set_docker_repository_config,
-)
 from exasol_integration_test_docker_environment.lib.docker.images.image_info import (
     ImageInfo,
 )
-from exasol_integration_test_docker_environment.lib.models.data.test_container_content_description import \
-    TestContainerContentDescription
+from exasol_integration_test_docker_environment.lib.models.config.build_config import (
+    set_build_config,
+)
+from exasol_integration_test_docker_environment.lib.models.config.docker_config import (
+    set_docker_repository_config,
+)
+from exasol_integration_test_docker_environment.lib.models.data.test_container_content_description import (
+    TestContainerContentDescription,
+)
 from exasol_integration_test_docker_environment.lib.test_environment.analyze_test_container import (
     AnalyzeTestContainer,
     DockerTestContainerBuild,
