+++ exasol_integration_test_docker_environment/lib/api/spawn_test_environment.py
@@ -17,21 +17,9 @@ from exasol_integration_test_docker_environment.cli.options.system_options impor
 from exasol_integration_test_docker_environment.cli.options.test_environment_options import (
     LATEST_DB_VERSION,
 )
-from exasol_integration_test_docker_environment.lib.models.api_errors import (
-    ArgumentConstraintError,
-)
 from exasol_integration_test_docker_environment.lib.base.run_task import (
     generate_root_task,
     run_task,
-)
-from exasol_integration_test_docker_environment.lib.config.build_config import (
-    set_build_config,
-)
-from exasol_integration_test_docker_environment.lib.config.docker_config import (
-    set_docker_repository_config,
-)
-from exasol_integration_test_docker_environment.lib.models.data.environment_info import (
-    EnvironmentInfo,
 )
 from exasol_integration_test_docker_environment.lib.docker.container.utils import (
     remove_docker_container,
@@ -41,6 +29,18 @@ from exasol_integration_test_docker_environment.lib.docker.networks.utils import
 )
 from exasol_integration_test_docker_environment.lib.docker.volumes.utils import (
     remove_docker_volumes,
+)
+from exasol_integration_test_docker_environment.lib.models.api_errors import (
+    ArgumentConstraintError,
+)
+from exasol_integration_test_docker_environment.lib.models.config.build_config import (
+    set_build_config,
+)
+from exasol_integration_test_docker_environment.lib.models.config.docker_config import (
+    set_docker_repository_config,
+)
+from exasol_integration_test_docker_environment.lib.models.data.environment_info import (
+    EnvironmentInfo,
 )
 from exasol_integration_test_docker_environment.lib.test_environment.parameter.docker_db_test_environment_parameter import (
     DbOsAccess,
