+++ exasol_integration_test_docker_environment/lib/base/db_os_executor.py
@@ -15,10 +15,10 @@ from docker.models.containers import (
 from paramiko.ssh_exception import NoValidConnectionsError
 
 from exasol_integration_test_docker_environment.lib.base.ssh_access import SshKey
+from exasol_integration_test_docker_environment.lib.docker import ContextDockerClient
 from exasol_integration_test_docker_environment.lib.models.data.database_info import (
     DatabaseInfo,
 )
-from exasol_integration_test_docker_environment.lib.docker import ContextDockerClient
 
 
 class DockerClientFactory:
