+++ exasol_integration_test_docker_environment/lib/docker/images/create/docker_image_build_task.py
@@ -5,12 +5,6 @@ from pathlib import Path
 from exasol_integration_test_docker_environment.lib.base.still_running_logger import (
     StillRunningLogger,
 )
-from exasol_integration_test_docker_environment.lib.config.build_config import (
-    build_config,
-)
-from exasol_integration_test_docker_environment.lib.config.docker_config import (
-    docker_build_arguments,
-)
 from exasol_integration_test_docker_environment.lib.docker.images.create.docker_image_creator_base_task import (
     DockerImageCreatorBaseTask,
 )
@@ -22,6 +16,12 @@ from exasol_integration_test_docker_environment.lib.docker.images.create.utils.b
 )
 from exasol_integration_test_docker_environment.lib.docker.images.image_info import (
     ImageInfo,
+)
+from exasol_integration_test_docker_environment.lib.models.config.build_config import (
+    build_config,
+)
+from exasol_integration_test_docker_environment.lib.models.config.docker_config import (
+    docker_build_arguments,
 )
