+++ exasol_integration_test_docker_environment/lib/docker/images/create/docker_image_load_task.py
@@ -1,10 +1,10 @@
 from pathlib import Path
 
-from exasol_integration_test_docker_environment.lib.config.build_config import (
-    build_config,
-)
 from exasol_integration_test_docker_environment.lib.docker.images.create.docker_image_creator_base_task import (
     DockerImageCreatorBaseTask,
+)
+from exasol_integration_test_docker_environment.lib.models.config.build_config import (
+    build_config,
 )
