+++ exasol_integration_test_docker_environment/lib/docker/images/push/push_log_handler.py
@@ -1,13 +1,13 @@
 import json
 
-from exasol_integration_test_docker_environment.lib.config.log_config import (
-    WriteLogFilesToConsole,
-)
 from exasol_integration_test_docker_environment.lib.docker.images.image_info import (
     ImageInfo,
 )
 from exasol_integration_test_docker_environment.lib.logging.abstract_log_handler import (
     AbstractLogHandler,
+)
+from exasol_integration_test_docker_environment.lib.models.config.log_config import (
+    WriteLogFilesToConsole,
 )
