+++ exasol_integration_test_docker_environment/lib/logging/abstract_log_handler.py
@@ -1,4 +1,6 @@
-from exasol_integration_test_docker_environment.lib.config.log_config import log_config
+from exasol_integration_test_docker_environment.lib.models.config.log_config import (
+    log_config,
+)
 
 
 class AbstractLogHandler:
