+++ exasol_integration_test_docker_environment/lib/models/config/build_config.py
