+++ exasol_integration_test_docker_environment/lib/models/config/docker_config.py
