+++ exasol_integration_test_docker_environment/lib/test_environment/spawn_test_database.py
@@ -32,6 +32,12 @@ from exasol_integration_test_docker_environment.lib.base.ssh_access import (
 from exasol_integration_test_docker_environment.lib.base.still_running_logger import (
     StillRunningLogger,
 )
+from exasol_integration_test_docker_environment.lib.docker.images.create.utils.pull_log_handler import (
+    PullLogHandler,
+)
+from exasol_integration_test_docker_environment.lib.docker.images.image_info import (
+    ImageInfo,
+)
 from exasol_integration_test_docker_environment.lib.models.data.container_info import (
     ContainerInfo,
 )
@@ -42,12 +48,6 @@ from exasol_integration_test_docker_environment.lib.models.data.docker_network_i
     DockerNetworkInfo,
 )
 from exasol_integration_test_docker_environment.lib.models.data.ssh_info import SshInfo
-from exasol_integration_test_docker_environment.lib.docker.images.create.utils.pull_log_handler import (
-    PullLogHandler,
-)
-from exasol_integration_test_docker_environment.lib.docker.images.image_info import (
-    ImageInfo,
-)
 from exasol_integration_test_docker_environment.lib.test_environment.db_version import (
     DbVersion,
 )
