+++ exasol_integration_test_docker_environment/test/test_api_logging.py
@@ -17,7 +17,7 @@ from exasol_integration_test_docker_environment.lib.base.run_task import (
     generate_root_task,
     run_task,
 )
-from exasol_integration_test_docker_environment.lib.config.build_config import (
+from exasol_integration_test_docker_environment.lib.models.config.build_config import (
     set_build_config,
 )
