+++ exasol_integration_test_docker_environment/test/test_base_task.py
@@ -8,14 +8,14 @@ from luigi import (
     Parameter,
 )
 
-from exasol_integration_test_docker_environment.lib.base.run_task import (
-    generate_root_task,
-)
 from exasol_integration_test_docker_environment.lib.base.dependency_logger_base_task import (
     DependencyLoggerBaseTask,
 )
 from exasol_integration_test_docker_environment.lib.base.json_pickle_parameter import (
     JsonPickleParameter,
+)
+from exasol_integration_test_docker_environment.lib.base.run_task import (
+    generate_root_task,
 )
 
 TestBaseTask = DependencyLoggerBaseTask
