+++ exasol_integration_test_docker_environment/test/test_generate_graph_plot.py
@@ -4,12 +4,12 @@ from pathlib import Path
 
 import luigi
 
+from exasol_integration_test_docker_environment.lib.base.dependency_logger_base_task import (
+    DependencyLoggerBaseTask,
+)
 from exasol_integration_test_docker_environment.lib.base.run_task import (
     generate_root_task,
     run_task,
-)
-from exasol_integration_test_docker_environment.lib.base.dependency_logger_base_task import (
-    DependencyLoggerBaseTask,
 )
