+++ exasol_integration_test_docker_environment/test/test_populate_data.py
@@ -10,10 +10,10 @@ import luigi
 from exasol_integration_test_docker_environment.lib.base.run_task import (
     generate_root_task,
 )
+from exasol_integration_test_docker_environment.lib.docker import ContextDockerClient
 from exasol_integration_test_docker_environment.lib.models.data.test_container_content_description import (
     TestContainerRuntimeMapping,
 )
-from exasol_integration_test_docker_environment.lib.docker import ContextDockerClient
 from exasol_integration_test_docker_environment.lib.test_environment.database_setup.populate_data import (
     PopulateTestDataToDatabase,
 )
