+++ test/integration/helpers.py
@@ -12,10 +12,10 @@ from exasol_integration_test_docker_environment.lib.base.db_os_executor import (
     DockerExecFactory,
     SshExecFactory,
 )
+from exasol_integration_test_docker_environment.lib.docker import ContextDockerClient
 from exasol_integration_test_docker_environment.lib.models.data.database_info import (
     DatabaseInfo,
 )
-from exasol_integration_test_docker_environment.lib.docker import ContextDockerClient
 from exasol_integration_test_docker_environment.lib.test_environment.parameter.docker_db_test_environment_parameter import (
     DbOsAccess,
 )
