+++ test/unit/test_termination_handler.py
@@ -14,12 +14,12 @@ import pytest
 from exasol_integration_test_docker_environment.cli.termination_handler import (
     TerminationHandler,
 )
+from exasol_integration_test_docker_environment.lib.base.dependency_logger_base_task import (
+    DependencyLoggerBaseTask,
+)
 from exasol_integration_test_docker_environment.lib.base.run_task import (
     generate_root_task,
     run_task,
-)
-from exasol_integration_test_docker_environment.lib.base.dependency_logger_base_task import (
-    DependencyLoggerBaseTask,
 )
